+++ src/columns.ts
@@ -29,7 +29,7 @@ type ColumnObj = {
   name: string
 }
 
-export abstract class Column<A = never, B = never> implements IStatementGiver {
+export abstract class Column implements IStatementGiver {
   private mTable?: Table
 
   protected constructor(protected readonly data: ColumnObj) {}
@@ -88,10 +88,6 @@ export abstract class Column<A = never, B = never> implements IStatementGiver {
     return new OrderByItemInfo(this, DESC, NULLS_LAST)
   }
 
-  public abstract eq(value: A): Condition
-
-  public abstract eq$(value: B): Condition
-
   public getStmt(data: BuilderData): string {
     if (this.mTable === undefined)
       throw new Error('Table of this column is undefined')
@@ -110,13 +106,13 @@ export abstract class Column<A = never, B = never> implements IStatementGiver {
   }
 }
 
-export class BooleanColumn extends Column<null|BooleanLike, null|boolean> implements Condition {
+export class BooleanColumn extends Column implements Condition {
   // START implement Condition
   public readonly leftExpression: Expression = new Expression(this)
   public readonly leftOperand: Operand = this.leftExpression.leftOperand
   public readonly type: ExpressionType = ExpressionType.BOOLEAN
 
-  public getColumns(): Column<null|BooleanLike, null|boolean>[] {
+  public getColumns(): BooleanColumn[] {
     return [this]
   }
   // END implement Condition
@@ -152,7 +148,7 @@ export class BooleanColumn extends Column<null|BooleanLike, null|boolean> implem
   }
 }
 
-export class NumberColumn extends Column<null|NumberLike, null|number> {
+export class NumberColumn extends Column {
   constructor(data: ColumnObj) {
     super(data)
   }
@@ -250,7 +246,7 @@ export class NumberColumn extends Column<null|NumberLike, null|number> {
   }
 }
 
-export class TextColumn extends Column<null|TextLike, null|string> {
+export class TextColumn extends Column {
   constructor(data: ColumnObj) {
     super(data)
   }