+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 
 ## [0.11.2]  2022-04-15
 ### Changes
-- create eq and eq$ function at parent class Column level
 - Deprecate PostgresBinder Type and getBinder() function
+- Add SelectItemInfo to select method
 
 ## [0.11.1]  2022-04-15
 ### Fixed